add forward_bg_val evaluation logic
good

+++ utils/utils.py
@@ -10,8 +10,10 @@ GUARANTEE_BACK_TH = 25000
 SENSE_LOW_BACK_TH = 20000
 LEN_SEQ_LOW = 16
 
+LEN_SEQ_NAIVE_BG = 8
 LEN_SEQ_NAIVE = 8
 ALARM_NAIVE_TH = 16
+ALARM_NAIVE_BG_LR = 1e-4
 
 LEN_SEQ = 16
 LEN_OVERLAP = 16
@@ -31,7 +33,7 @@ ALARM_LOW_CNT_TH_SVM = 5
 ALARM_LOW_CNT_DECAY = 0.1
 ALARM_LOW_NEG_SCORE_WEIGHT = 2
 ALARM_LOW_SMOOTH_TH = 1000
-DEBUG_ALARM_INDICATOR_VAL = 2 ** 11
+DEBUG_ALARM_INDICATOR_VAL = 2 ** 10
 ALARM_LOW_ANCHOR_STEP = 2
 
 
@@ -167,8 +169,8 @@ def find_key_idx(seq, th_val=10, th_cnt=10, th_delta=10):
     return key_idx if flag_valid else -1
 
 
-def make_dirs(dir_root):
-    if os.path.exists(dir_root):
+def make_dirs(dir_root, reset=False):
+    if os.path.exists(dir_root) and reset:
         shutil.rmtree(dir_root)
     os.makedirs(os.path.join(dir_root), exist_ok=True)
